works

+++ transaction.go
@@ -4,6 +4,6 @@ type Transaction struct {
 	EventId   string                 `json:"event_id"`
 	Release   string                 `json:"release"`
 	User      map[string]interface{} `json:"user"`
-	Timestamp int64                  `json:"timestamp"`
+	Timestamp float64                `json:"timestamp"`
 	// Type      string                 `json:"type"`
 }
